Make tests independent of the details of letter spacing
There are plenty of test checking whether letter spacing adds a space at
the end of the line or not. Confirming that is not the point of these
two tests, especially given that that is still a controvertial topic,
while the point of this test is not.

This commit makes the tests work regardless of how end of lines are
handled.